requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
audioop-lts==0.2.2
autobahn==24.4.2
Automat==25.4.16
av==15.1.0
beaker-py==2.4.6
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.39.6
botocore==1.39.6
CacheControl==0.14.3
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.3.0
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
cryptography==45.0.5
ctranslate2==4.6.0
daphne==4.2.1
Deprecated==1.2.18
Django==5.2.4
etelemetry==0.3.1
faster-whisper==1.2.0
filelock==3.18.0
firebase-admin==6.9.0
fitz==0.0.1.dev2
flatbuffers==25.2.10
fsspec==2025.7.0
ftfy==6.3.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
img2pdf==0.6.1
incremental==24.7.2
Jinja2==3.1.6
jmespath==1.0.1
lingua-language-detector==2.1.1
llvmlite==0.44.0
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==1.1.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numba==0.61.2
numpy==1.26.4
ollama==0.5.4
onnxruntime==1.22.1
openai-whisper==20250625
OpenCC==1.1.9
orjson==3.11.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pikepdf==9.10.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.8.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.2.1
redis==6.4.0
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scipy==1.16.2
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.2
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
wrapt==1.17.2
zope.interface==7.2
zstandard==0.23.0
